udpate

+++ model/bert.py
@@ -14,7 +14,7 @@ class BERTNER(nn.Module):
         
     def forward(self, ids, masks, labels):
         _, logits = self.emission(input_ids=ids, attention_mask=masks, labels=labels).to_tuple()
-        loss = -self.crf(logits, labels, masks)
+        loss = -self.crf(logits, labels, torch.ByteTensor(masks))
         logits = self.crf.decode(logits)
         return loss, logits
